further improve flasher UX by printing version number before flash
