+++ proscli/utils.py
@@ -1,8 +1,25 @@
+import os
+
 import click
+import sys
+
 from proscli.state import State
 
 pass_state = click.make_pass_decorator(State)
 
+def get_version():
+    try:
+        if os.path.isfile(os.path.join(__file__, '../../version')):
+            return open(os.path.join(__file__, '../../version')).read().strip()
+    except Exception:
+        pass
+    try:
+        if getattr(sys, 'frozen', False):
+            import BUILD_CONSTANTS
+            return BUILD_CONSTANTS.CLI_VERSION
+    except Exception:
+        pass
+    return None # Let click figure it out
 
 def verbosity_option(f):
     """
